+++ sdmx-dl-desktop/src/main/java/sdmxdl/desktop/MainComponent.java
@@ -188,7 +188,7 @@ public final class MainComponent extends JComponent implements HasSdmxProperties
         main.putClientProperty(FlatClientProperties.TABBED_PANE_TAB_TYPE, FlatClientProperties.TABBED_PANE_TAB_TYPE_CARD);
 
         sourcesList.setCellRenderer(JLists.cellRendererOf((label, value) -> {
-            label.setText(nameDescription(value.getId(), value.getDescription(LanguagePriorityList.ANY)).render());
+            label.setText(nameDescription(value.getId(), value.getName(LanguagePriorityList.ANY)).render());
             label.setIcon(getDataSourceIcon(value.getId(), 24, sourcesList::repaint));
             label.setBorder(BorderFactory.createEmptyBorder(3, 3, 3, 3));
         }));