+++ sdmx-dl-provider-connectors/src/main/java/internal/sdmxdl/provider/connectors/drivers/OecdDriver.java
@@ -46,7 +46,7 @@ public final class OecdDriver implements WebDriver {
             .source(SdmxWebSource
                     .builder()
                     .id("OECD")
-                    .descriptionOf("The Organisation for Economic Co-operation and Development")
+                    .nameOf("The Organisation for Economic Co-operation and Development")
                     .driver(CONNECTORS_OECD)
                     .endpointOf("https://stats.oecd.org/restsdmx/sdmx.ashx")
                     .websiteOf("https://stats.oecd.org")