+++ sdmx-dl-provider-connectors/src/main/java/internal/sdmxdl/provider/connectors/drivers/UisDriver.java
@@ -51,7 +51,7 @@ public final class UisDriver implements WebDriver {
             .source(SdmxWebSource
                     .builder()
                     .id("UIS")
-                    .descriptionOf("Unesco Institute for Statistics")
+                    .nameOf("Unesco Institute for Statistics")
                     .driver(CONNECTORS_UIS)
                     .endpointOf(FALLBACK_ENDPOINT)
                     .websiteOf("http://data.uis.unesco.org")