+++ sdmx-dl-provider-ri/src/main/java/internal/sdmxdl/provider/ri/web/drivers/ImfDriver2.java
@@ -66,8 +66,8 @@ public final class ImfDriver2 implements WebDriver {
             .source(SdmxWebSource
                     .builder()
                     .id("IMF")
-                    .descriptionOf("International Monetary Fund")
-                    .description("en", "International Monetary Fund")
+                    .nameOf("International Monetary Fund")
+                    .name("en", "International Monetary Fund")
                     .driver(RI_IMF)
                     .endpointOf("http://dataservices.imf.org/REST/SDMX_XML.svc")
                     .websiteOf("https://data.imf.org")