+++ sdmx-dl-provider-ri/src/main/java/internal/sdmxdl/provider/ri/web/drivers/RngDriver.java
@@ -51,8 +51,8 @@ public final class RngDriver implements WebDriver {
             .source(SdmxWebSource
                     .builder()
                     .id("RNG")
-                    .descriptionOf("Random number generator")
-                    .description("en", "Random number generator")
+                    .nameOf("Random number generator")
+                    .name("en", "Random number generator")
                     .driver(RI_RNG)
                     .endpointOf("rng:3:4:0:2010-01-01")
                     .build())